refactor: rename omni-key files to mpc-key
- Rename src/omni-key.ts to src/mpc-key.ts
- Rename tests/unit/omni-key.test.ts to tests/unit/mpc-key.test.ts
- Update all import paths to use new filename
- Filename now aligns with MPCKey/MockMPCKey class names
